requirements.txt
flask
flask_login
paramiko
scp
pyyaml
loguru
